+++ src/main/java/com/pss/premierservicesolutions/repositories/ComplaintRepository.java
@@ -1,9 +1,9 @@
 package com.pss.premierservicesolutions.repositories;
 
-import com.pss.premierservicesolutions.models.Complaint;
+import com.pss.premierservicesolutions.entity.Complaint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComplaintRepository extends JpaRepository<Complaint, Integer> {
+public interface ComplaintRepository extends JpaRepository<Complaint, Long> {
 }
