+++ src/main/java/com/pss/premierservicesolutions/repositories/ContractRepository.java
@@ -1,12 +1,12 @@
 package com.pss.premierservicesolutions.repositories;
 
-import com.pss.premierservicesolutions.models.Contract;
+import com.pss.premierservicesolutions.entity.Contract;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long> {
     List<Contract> findAllById(long clientId);
 }
