+++ src/main/java/com/pss/premierservicesolutions/repositories/EmployeeRepository.java
@@ -1,9 +1,9 @@
 package com.pss.premierservicesolutions.repositories;
 
-import com.pss.premierservicesolutions.models.Employee;
+import com.pss.premierservicesolutions.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
