+++ src/main/java/com/pss/premierservicesolutions/repositories/ServiceLevelAgreementRepository.java
@@ -1,9 +1,9 @@
 package com.pss.premierservicesolutions.repositories;
 
-import com.pss.premierservicesolutions.models.ServiceLevelAgreement;
+import com.pss.premierservicesolutions.entity.ServiceLevelAgreement;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServiceLevelAgreementRepository extends JpaRepository<ServiceLevelAgreement, Integer> {
+public interface ServiceLevelAgreementRepository extends JpaRepository<ServiceLevelAgreement, Long> {
 }
